cache ttl simplification + cache prefixes separated by domains
