--- packages/api/src/cache/enums/cache-prefixes.enum.ts
@@ -1,4 +1,0 @@
-export enum CachePrefixes {
-  CURRENT_WEATHER = 'currentWeather:',
-  CITY_VALIDATION = 'cityValidation:',
-}
