--- packages/api/src/cache/enums/cache-ttl.enum.ts
@@ -1,7 +1,0 @@
-import { HOUR, MINUTE } from 'src/common/utils/time-units';
-
-export enum CacheTTL {
-  MINUTES_10 = 10 * MINUTE,
-  MINUTES_15 = 15 * MINUTE,
-  HOUR_1 = HOUR,
-}
