+++ packages/api/src/city/constants/enums/city-cache-prefixes.enum.ts
@@ -1,0 +1,3 @@
+export enum CityCachePrefixes {
+  CITY_VALIDATION = 'cityValidation:',
+}
