+++ packages/api/src/city/constants/types/city-fetch.type.ts
