+++ packages/api/src/common/proxies/city-validation-cache.proxy.ts
@@ -1,13 +1,10 @@
-import { CachePrefixes } from 'src/cache/enums/cache-prefixes.enum';
-import { CacheTTL } from 'src/cache/enums/cache-ttl.enum';
 import { CacheAccessor } from 'src/cache/interfaces/cache-service.interface';
 import { transformKey } from 'src/cache/utils/key-transformation';
+import { CityCachePrefixes } from 'src/city/constants/enums/city-cache-prefixes.enum';
 import { ChainableCityProvider } from 'src/city/interfaces/chainable-city.provider';
+import { HOUR } from '../utils/time-units';
 
 export class CityProviderCacheProxy extends ChainableCityProvider {
-  private readonly keyBase: CachePrefixes = CachePrefixes.CITY_VALIDATION;
-  private readonly ttl: CacheTTL = CacheTTL.HOUR_1;
-
   constructor(
     private readonly wrapped: ChainableCityProvider,
     private readonly cacheService: CacheAccessor,
@@ -16,12 +13,12 @@ export class CityProviderCacheProxy extends ChainableCityProvider {
   }
 
   async validateCity(city: string): Promise<boolean> {
-    const cacheKey = transformKey(this.keyBase, city);
+    const cacheKey = transformKey(CityCachePrefixes.CITY_VALIDATION, city);
     const cached = await this.cacheService.get<boolean>(cacheKey);
     if (cached) return cached; // cache hit
 
     const result: boolean = await this.wrapped.handle(city);
-    await this.cacheService.set<boolean>(cacheKey, result, this.ttl); // cache validation result
+    await this.cacheService.set<boolean>(cacheKey, result, HOUR); // cache validation result
 
     return result;
   }
