+++ packages/api/src/weather/constants/enums/weather-cache-prefixes.enum.ts
@@ -1,0 +1,3 @@
+export enum WeatherCachePrefixes {
+  CURRENT_WEATHER = 'currentWeather:',
+}
